[#153] Implement a PR flow on CLI (#184)
* [#156] Create a PR item on home and a menu view (#157)
* [#158] CLI implementation for the PR flow (#161)
* [#159] Resources templates improvements and modifications (#162)
* [#160] Models and schemas modifications for CLI PR flow (#163)
* [#164] Treat the errors on the PR flow (#165)
* [#166] Look for bugs and fix them (#173)
* [#167] Create alias for the directories (#168)
* [#169] Implement the current task version approach (#170)
* [#171] Improve the approach on templates (#172)
* [#166] Fixing bugs found
* [#153] Fixing small issue
* [#175] Implement a Screen CLI helper (#176)
* [#174] Implement a file change description flow (#177)
* [#175] Implement a Screen CLI helper
* [#174] Implement a file change description flow
* [#178] Create a custom PR template to use in general cases
* {#181] Fix issues caused by last PR
* [#180] Implement same approach as on PR to commit file description

+++ src/services/FS/index.js
@@ -2,6 +2,20 @@ const fs = require('fs');
 const path = require('path');
 
 class FileSystemService {
+    static readFileSync(path) {
+        return fs.readFileSync(path, { encoding: 'utf-8' });
+    }
+
+    static async readFile(path) {
+        return new Promise((resolve, reject) => {
+            fs.readFile(path, (err, dataBuff) => {
+                if (err) throw reject(new Error.Log(err));
+                
+                return resolve(dataBuff);
+            });
+        });
+    }
+
     static async copyFiles(filesToCopy, sourceDir, destDir) {
         if (!Array.isArray(filesToCopy)) {
             throw new Error.Log({
@@ -33,12 +47,16 @@ class FileSystemService {
                         fs.mkdirSync(newDestDir, { recursive: true });
                     }
 
-                    fs.copyFile(sourcePath, destPath, (err) => {
-                        if (err) throw reject(new Error.Log(err));
-        
-                        console.log(`${file} was copied to ${destPath}`);
+                    if (fs.existsSync(sourcePath)) {
+                        fs.copyFile(sourcePath, destPath, (err) => {
+                            if (err) throw reject(new Error.Log(err));
+            
+                            console.log(`${file} was copied to ${destPath}`);
+                            return resolve(file.filename);
+                        });
+                    } else {
                         return resolve(file.filename);
-                    });
+                    }
                 }));
             });
